Move server.test & seed, increase reviews per listing by tenfold, updated npm scripts (not remotely relevent to this feature!)
